Update to MorphSearch.cs: A Sort Button!
Often is needed to know which morphs've different values than 0, so I added a button which switch between two sortings methods:by name and by how far from 0 the morph is (by Value). But I'm not a C# programmer so I've done some logic tricks to get the job done.